requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
PyMuPDF>=1.24.0
reportlab>=4.1.0
pydantic>=2.5.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
Pillow>=10.0.0
